made RPC posting a member function RPC client
